requirements.txt
streamlit==1.32.0
tensorflow==2.16.1
scikit-learn==1.4.1
xgboost==2.0.3
opencv-python==4.9.0.80
pandas==2.2.1
numpy==1.26.4
joblib==1.3.2
Pillow==10.2.0
matplotlib==3.8.3
